+++ packages/xchain-thorchain-query/__tests__/decimal-math.test.ts
@@ -1,4 +1,3 @@
-import { Network } from '@xchainjs/xchain-client'
 import {
   // Asset,
   // AssetBNB,
@@ -15,14 +14,9 @@ import {
 import mockMidgardApi from '../__mocks__/midgard-api'
 import mockThornodeApi from '../__mocks__/thornode-api'
 import { CryptoAmount } from '../src/crypto-amount'
-import { ThorchainCache } from '../src/thorchain-cache'
 import { ThorchainQuery } from '../src/thorchain-query'
-import { Midgard } from '../src/utils/midgard'
-import { Thornode } from '../src/utils/thornode'
 
-const thorchainCache = new ThorchainCache(new Midgard(Network.Mainnet), new Thornode(Network.Mainnet))
-
-const thorchainQuery = new ThorchainQuery(thorchainCache)
+const thorchainQuery = new ThorchainQuery()
 
 const assetUSDC = assetFromStringEx('ETH.USDC-0XA0B86991C6218B36C1D19D4A2E9EB0CE3606EB48')
 // const BUSD = assetFromString('BNB.BUSD-BD1')
@@ -87,4 +81,12 @@ describe('decimal math tests', () => {
     expect(amount.assetAmountFixedString()).toEqual(expected.assetAmountFixedString())
     expect(amount.eq(expected)).toBe(true)
   })
+  // it(`Should do decimal conversion `, async () => {
+  //   const input = new CryptoAmount(assetToBase(assetAmount(100)), AssetRuneNative)
+  //   const out = getBaseAmountWithDiffDecimals(input.baseAmount, 18)
+  //   const amount = await thorchainQuery.thorchainCache.convert(input, AssetBTC)
+  //   const expected = new CryptoAmount(assetToBase(assetAmount(`0.00917714`)), AssetBTC)
+  //   expect(amount.assetAmountFixedString()).toEqual(expected.assetAmountFixedString())
+  //   expect(amount.eq(expected)).toBe(true)
+  // })
 })