+++ packages/xchain-thorchain-query/__tests__/thorchain-query.test.ts
@@ -1,4 +1,5 @@
 import {
+  AssetAtom,
   AssetBNB,
   AssetBTC,
   AssetETH,
@@ -19,7 +20,8 @@ import { EstimateSwapParams, TxDetails } from '../src/types'
 const thorchainQuery = new ThorchainQuery()
 
 const assetUOS = assetFromStringEx('ETH.UOS-0XD13C7342E1EF687C5AD21B27C2B65D772CAB5C8C')
-const assetUSDC = assetFromStringEx('ETH.USDC-0XA0B86991C6218B36C1D19D4A2E9EB0CE3606EB48')
+const assetEthUSDC = assetFromStringEx('ETH.USDC-0XA0B86991C6218B36C1D19D4A2E9EB0CE3606EB48')
+const assetAVAXUSDC = assetFromStringEx(`AVAX.USDC-0XB97EF9EF8734C71904D8002F8B6BC66DD9C48A6E`)
 const BUSD = assetFromStringEx('BNB.BUSD-BD1')
 
 function printTx(txDetails: TxDetails, input: CryptoAmount) {
@@ -71,7 +73,7 @@ describe('Thorchain-query tests', () => {
 
   it('Should fail estimate swap because destination chain is halted ', async () => {
     const swapParams: EstimateSwapParams = {
-      input: new CryptoAmount(assetToBase(assetAmount(2)), AssetETH),
+      input: new CryptoAmount(assetToBase(assetAmount(2, 18)), AssetETH),
       destinationAsset: AssetLTC,
       destinationAddress: 'xxx',
     }
@@ -101,7 +103,7 @@ describe('Thorchain-query tests', () => {
   })
   it('Should estimate swap from USDC to RUNE ', async () => {
     const swapParams: EstimateSwapParams = {
-      input: new CryptoAmount(assetToBase(assetAmount(1000, 6)), assetUSDC),
+      input: new CryptoAmount(assetToBase(assetAmount(1000, 6)), assetEthUSDC),
       destinationAsset: AssetRuneNative,
       destinationAddress: 'xxx',
     }
@@ -111,13 +113,13 @@ describe('Thorchain-query tests', () => {
     expect(estimate.txEstimate.canSwap).toEqual(true)
     expect(estimate.txEstimate.netOutput.assetAmount.decimal).toEqual(8)
     expect(estimate.txEstimate.netOutput.assetAmount.amount().toFixed()).toEqual(
-      assetAmount('499.68215078').amount().toFixed(),
+      assetAmount('499.62215078').amount().toFixed(),
     )
   })
   it('Should estimate swap from RUNE to USDC ', async () => {
     const swapParams: EstimateSwapParams = {
       input: new CryptoAmount(assetToBase(assetAmount(50)), AssetRuneNative),
-      destinationAsset: assetUSDC,
+      destinationAsset: assetEthUSDC,
       destinationAddress: 'runeaddress',
     }
 
@@ -126,7 +128,7 @@ describe('Thorchain-query tests', () => {
     expect(estimate.txEstimate.canSwap).toEqual(true)
     expect(estimate.txEstimate.netOutput.assetAmount.decimal).toEqual(6)
     expect(estimate.txEstimate.netOutput.assetAmount.amount().toFixed()).toEqual(
-      assetAmount('99.818529').amount().toFixed(),
+      assetAmount('96.181146').amount().toFixed(),
     )
   })
   it('Should estimate swap from BUSD to RUNE ', async () => {
@@ -141,7 +143,7 @@ describe('Thorchain-query tests', () => {
     printTx(estimate, swapParams.input)
     expect(estimate.txEstimate.canSwap).toEqual(true)
     expect(estimate.txEstimate.netOutput.assetAmount.amount().toFixed()).toEqual(
-      assetAmount('500.0078375').amount().toFixed(),
+      assetAmount('499.9478375').amount().toFixed(),
     )
   })
   it('Should estimate swap from UOS to ETH ', async () => {
@@ -156,7 +158,7 @@ describe('Thorchain-query tests', () => {
     expect(estimate.txEstimate.canSwap).toEqual(true)
     // expect(estimate.txEstimate.netOutput.assetAmount.decimal).toEqual(18)
     expect(estimate.txEstimate.netOutput.assetAmount.amount().toFixed()).toEqual(
-      assetAmount('1.27226837').amount().toFixed(),
+      assetAmount('1.28763837').amount().toFixed(),
     )
   })
   it('Should estimate swap from ETH to UOS ', async () => {
@@ -171,7 +173,35 @@ describe('Thorchain-query tests', () => {
     expect(estimate.txEstimate.canSwap).toEqual(true)
     expect(estimate.txEstimate.netOutput.assetAmount.decimal).toEqual(4)
     expect(estimate.txEstimate.netOutput.assetAmount.amount().toFixed()).toEqual(
-      assetAmount('3802.1387').amount().toFixed(),
+      assetAmount('3794.0994').amount().toFixed(),
     )
+  })
+  it('Should construct the correct memo BUSD->USDC', async () => {
+    const swapParams: EstimateSwapParams = {
+      input: new CryptoAmount(assetToBase(assetAmount(100)), BUSD),
+      destinationAsset: assetEthUSDC,
+      destinationAddress: 'xxx',
+    }
+    const estimate = await thorchainQuery.estimateSwap(swapParams)
+    const correctMemo = `=:ETH.USDC-0XA0B86991C6218B36C1D19D4A2E9EB0CE3606EB48:xxx:9621201999`
+    expect(estimate.memo).toEqual(correctMemo)
+  })
+  it('Should construct the correct memo ETH->USDC', async () => {
+    const swapParams: EstimateSwapParams = {
+      input: new CryptoAmount(assetToBase(assetAmount(1, 18)), AssetETH),
+      destinationAsset: assetEthUSDC,
+      destinationAddress: 'xxx',
+    }
+    const estimate = await thorchainQuery.estimateSwap(swapParams)
+    const correctMemo = `=:ETH.USDC-0XA0B86991C6218B36C1D19D4A2E9EB0CE3606EB48:xxx:173329271999`
+    expect(estimate.memo).toEqual(correctMemo)
+  })
+  it(`Should check assets match asset pools`, async () => {
+    const assetPoolEthUsdc = await thorchainQuery.thorchainCache.getPoolForAsset(assetEthUSDC)
+    const assetPoolAvaxUsdc = await thorchainQuery.thorchainCache.getPoolForAsset(assetAVAXUSDC)
+    const assetPoolGai = await thorchainQuery.thorchainCache.getPoolForAsset(AssetAtom)
+    expect(assetPoolEthUsdc.asset).toEqual(assetEthUSDC)
+    expect(assetPoolAvaxUsdc.asset).toEqual(assetAVAXUSDC)
+    expect(assetPoolGai.asset).toEqual(AssetAtom)
   })
 })