+++ packages/xchain-midgard/CHANGELOG.md
@@ -1,3 +1,14 @@
+# v0.0.2.0 (2022-10-18)
+
+## breaking changes
+
+- constructor defaults to 9R endpoint
+- removed https://midgard.thorchain.info/ endpoint
+
+# Add
+
+- new bindings generated for midgard 2.9.5
+
 # v0.0.1.0 (2022-10-04)
 
 ## Update