+++ packages/xchain-thorchain-query/CHANGELOG.md
@@ -1,8 +1,22 @@
+# v0.1.4 (2022-10-17)
+
+## Fix
+
+- Bug - Limit asset amount was not using 8 decimal places 
+- Bug - outbound fee is now calculated correctly
+- Bug - affiliate fee calculation fixed 
+- Bug - getPoolForAsset() changed to take both chain & ticker for lookup
+- Bump dep on package `xchain-midgard`
+
+## Add
+
+- check input asset decimals match nativeDecimals in LiquidityPool.pool, or throw Error
+
 # v0.1.3 (2022-10-17)
 
 ## Update
 
-- default to mainnet and stadard APIs with no arg constructor 
+- default to mainnet and standard APIs with no arg constructor
 
 # v0.1.2 (2022-10-11)